# ent/foodmenu.go
// Code generated by entc, DO NOT EDIT.

package ent

import (
	"fmt"
	"strings"

	"github.com/facebook/ent/dialect/sql"
	"github.com/faomg201/app/ent/foodmenu"
)

// FOODMENU is the model entity for the FOODMENU schema.
type FOODMENU struct {
	config `json:"-"`
	// ID of the ent.
	ID int `json:"id,omitempty"`
	// FOODMENUNAME holds the value of the "FOODMENU_NAME" field.
	FOODMENUNAME string `json:"FOODMENU_NAME,omitempty"`
	// Edges holds the relations/edges for other nodes in the graph.
	// The values are being populated by the FOODMENUQuery when eager-loading is set.
	Edges FOODMENUEdges `json:"edges"`
}

// FOODMENUEdges holds the relations/edges for other nodes in the graph.
type FOODMENUEdges struct {
	// FOODMENURECORD holds the value of the FOODMENU_RECORD edge.
	FOODMENURECORD []*Recordfood
	// loadedTypes holds the information for reporting if a
	// type was loaded (or requested) in eager-loading or not.
	loadedTypes [1]bool
}

// FOODMENURECORDOrErr returns the FOODMENURECORD value or an error if the edge
// was not loaded in eager-loading.
func (e FOODMENUEdges) FOODMENURECORDOrErr() ([]*Recordfood, error) {
	if e.loadedTypes[0] {
		return e.FOODMENURECORD, nil
	}
	return nil, &NotLoadedError{edge: "FOODMENU_RECORD"}
}

// scanValues returns the types for scanning values from sql.Rows.
func (*FOODMENU) scanValues() []interface{} {
	return []interface{}{
		&sql.NullInt64{},  // id
		&sql.NullString{}, // FOODMENU_NAME
	}
}

// assignValues assigns the values that were returned from sql.Rows (after scanning)
// to the FOODMENU fields.
func (f *FOODMENU) assignValues(values ...interface{}) error {
	if m, n := len(values), len(foodmenu.Columns); m < n {
		return fmt.Errorf("mismatch number of scan values: %d != %d", m, n)
	}
	value, ok := values[0].(*sql.NullInt64)
	if !ok {
		return fmt.Errorf("unexpected type %T for field id", value)
	}
	f.ID = int(value.Int64)
	values = values[1:]
	if value, ok := values[0].(*sql.NullString); !ok {
		return fmt.Errorf("unexpected type %T for field FOODMENU_NAME", values[0])
	} else if value.Valid {
		f.FOODMENUNAME = value.String
	}
	return nil
}

// QueryFOODMENURECORD queries the FOODMENU_RECORD edge of the FOODMENU.
func (f *FOODMENU) QueryFOODMENURECORD() *RecordfoodQuery {
	return (&FOODMENUClient{config: f.config}).QueryFOODMENURECORD(f)
}

// Update returns a builder for updating this FOODMENU.
// Note that, you need to call FOODMENU.Unwrap() before calling this method, if this FOODMENU
// was returned from a transaction, and the transaction was committed or rolled back.
func (f *FOODMENU) Update() *FOODMENUUpdateOne {
	return (&FOODMENUClient{config: f.config}).UpdateOne(f)
}

// Unwrap unwraps the entity that was returned from a transaction after it was closed,
// so that all next queries will be executed through the driver which created the transaction.
func (f *FOODMENU) Unwrap() *FOODMENU {
	tx, ok := f.config.driver.(*txDriver)
	if !ok {
		panic("ent: FOODMENU is not a transactional entity")
	}
	f.config.driver = tx.drv
	return f
}

// String implements the fmt.Stringer.
func (f *FOODMENU) String() string {
	var builder strings.Builder
	builder.WriteString("FOODMENU(")
	builder.WriteString(fmt.Sprintf("id=%v", f.ID))
	builder.WriteString(", FOODMENU_NAME=")
	builder.WriteString(f.FOODMENUNAME)
	builder.WriteByte(')')
	return builder.String()
}

// FOODMENUs is a parsable slice of FOODMENU.
type FOODMENUs []*FOODMENU

func (f FOODMENUs) config(cfg config) {
	for _i := range f {
		f[_i].config = cfg
	}
}